cleanup: add propper logging
this adds a --loglevel flag which sets the global logrus logger. The
rest of the code then should use logrus' functions like logrus.Info().
With this, we can remove the rest of the hand rolled debug
infrastructrue.

Again, try to be as minimal inversive as possible. There needs to be a
lot of further cleanup.