add controller variable
So far this merely worked by conincidence (i.e., when executed by on the
node that also is the LINSTOR controller, because of some "localhost
fallback" in golinstor). Add --controller.

Again, his adds even one additional flag to the monstrous functions
that already take close to like 10 params until we finally structure the
code properly.