cleanup: introduce a Linstor type
This gets rid of the first part of these functions that take ways too
many parameters. Provide this configuration type with the according
methods.

We could have NewCreate/DeleteConfig(...) *Linstor helpers, but so far
the overall Linstor config struct is simple enough.

This also fixes a bug where the error of Delete was actually ignored.