cleanup: introduce iscsi.Target
This type contains the iSCSI target configuration. Again this is simple
enough, no 'NewTarget(...) *Target' needed. Here it actually might be
nice, as we could keep all fields private and do the target name
extraction and check at New time. Consider this for the future, there is
too much work to get this in shape to care about it right now.

While at it, rework the target extraction method, no need to C-style
handroll this.