cleanup: switch to go templates
This switches to Go templates, no need to hand roll this.

As I don't expect users to modify templates, we can use consts and
therefore embedded them in the binary. With this we only need to ship the
single binary. Let's see if there are situations where we would want
them to be modifiable. If so, this part is trivial to change.

Fix some programming errors on the way (e.g, ignored errs).

This is again the minimal inversive change, there is a lot that needs
cleaning in this area.