cleanup: simplify extcmd
I don't see a reason at all why this hand rolls command execution on the
lowest level possible. It is all there in the os/exec stdlib.

While at it, switch from "lines" to strings. We don't need lines. We get
stuff (most likely xml), which we then parse as a string blob. Code that
needs to do that, can use a Scanner. If we ever need that, we can add a
wrapper to extcmd, but currently we don't have a single use case and it
just complicates things (see the FuseStrings that got kicked out).