cleanup: rm the ElementIdxSet
Again, this can be trivially implemented with standard maps. It removes
a whole lot of external code we dependend on. As a side note: It took me
some time to actually figure out the reverse nature of this thing, as it
is completely hidden in the NewElemIdxSet() constructor, while
ReverseSortedKeys() should transport this message.

While not directly visible, this removes hundreds LOCs (the godsaext
library has 766 lines as of writing this, not sure how much it actually
used). A library that has exactly 0 upstream tests. Sorry, but I don't
want to rely on this for no reason. This replaces it with 20 lines
(this and the previous commit; I counted them) of standard Go.

Speaking of testing most of it needs a rewrite to even allow for testing
(e.g., taking io.Readers,...)