cleanup: merge the two sets into one IntSet, rm numberpool
This merges the two sets (one int one and one int16 one). It also
removes the number pool generator. I expecially don't
like that it hides the fact where it actually starts, but in general it
contains no tests and is not "obviously correct". There is no reason to
make this harder to maintain.

Iterating the keys to find a new free ID is O(n), but it is trivial to
see that the code is correct, and it contains tests. The worst case (see
go test -bench) is handled in less than a 1ms, I guess we don't
need premature optimisation at this point. Especially when the actual
worst case is determined by external command execution and cluster
operations that take magnitudes longer.