crm,iscsi,list: overhaul CRM config parsing
In order to make a better list command which can display a sensible
status summary, we need to actually get that data first.
As the original CrmConfiguration data structure had pretty limited
information, enhance the parsing process first.

Also extend iscsi.Target to have multiple LUNs, since that's more in
line with how iSCSI actually does things. Some operations have also been
adapted to run multiple times for all LUs if necessary.

This also introduces a rough framework of what the output of the list
command is supposed to look like in the end, with the table and all.