let LUNs have a size
On a higher level view it is the LUNs that have a size, so add this
field to the struct. This resolves the current issue with sizes as
follows:

- the higher level Linstor struct still has a size, it can be used with
a size to create a LINSTOR resource. The Linstor (only) interface still
makes sense.
- the ISCSI part sets the size in the Linstor struct like it does with
the name according to the actual LUN sizes.
- The cli only sets one LUN.
- We can have multiple LUNs via the REST API. The REST API doc will not
even mention the Linstor.SizeKiB, users are supposed to set the
LUN.SizeKiB field.