all: do not pass maps as pointers
A map in go is already (kind of) a pointer, so there is no need to pass
it as a pointer. The only use case would be that the whole map needs to
be re-allocated and replaced with another map, but that never happens
here. The maps are only ever modified, which is safe to do when passing
the map "per value".