introduce TargetConfig
We want a way to check if the IQN is valid. We need it for the CLI, but
also for the REST interface. The most idiomatic way is to provide a
"config struct", and provide New* functions. Currently the Target does
not contain state on its own, so it just embedds the TargetConfig.

One advantage is that we can remove unused cobra functions, but this is
mainly a preparation for the REST-API.