crmcontrol: make CRM object map a slice
The CRM object map was used as a Set, but it was mostly iterated over
anyways. For the few times where we do need random access, performance
really doesn't matter (we're talking at most 10 entries in this slice).

On the up side, this paves the road for changing the data structure of
the "Probe" command.