list: introduce ResourceRunState
Saving the resource states in a map keyed with their name is really
dynamic, but it makes actually handling and displaying the state data
needlessly complex.

Thus, introduce a new data structure to store all state related
information about a CRM resource and pass it around.

Note: this breaks the "probe" command, but since it will soon be
completely removed anyways, just stub it out for now.