iscsi: compile IQN regex beforehand
Since the strings used to build the regex are constants, the compilation
can never really go wrong (except during development, where a panic is
fine). So we can get some complexity out of the CheckIQN function by not
having to check whether or not the regex could be compiled every time.