REST: introduce global mutex
For now I'm pretty sure we are far from being concurrency safe. I'm
pretty sure this will never be a problem in this case. I guess we would
have to serialize CRM/Pacemaker actions anyways, so there would not be
too much gain. Anyways, fix this with a mutex.

We could create a lockfile for plain client use, there is one good
implementation, but for now I think it is just relevant for the
REST-API.