delete: try to delete linstor part even if CRM part failed
So far we exited early if for example the resource did not exist in
pacemaker anymore. I think this is an error, something went wrong, and
we should report it as error. But we just can try to delete the LINSTOR
resource. This makes it easier for the user to clean up manually, and it
might avoid a later fallout where an undeleted LINSTOR resource already
exists at a later create operation.