Do not return the XML document from modifyCrmTargetRole
The function returned either the pointer that was passed to it
or nil, but that did not seem to serve any purpose.
Most callers call the function repeatedly, and if those callers
replace their XML document pointer with nil, they will call
the modifyCrmTargetRole function again with a nil pointer,
which causes a crash.