crmcontrol: factor out low-level API
Package crmcontrol is supposed to be a low-level interface to the CIB.
However, until now it contained a lot of iSCSI-specific code. This makes
it hard to reuse the code in other projects (for example
linstor-appliance, which was the motivation for this change).

This splits some of the low-level functionality of crmcontrol into a
completely separate project, which will be located at github.com/LINBIT/crm.

In theory, all iSCSI-related code should be left in this package, so
that package crm can be reused at will.