linstor: consider diskless resources good
For every resource, we check that its disk state is "UpToDate". This is
a pretty basic check, and as it turns out, it was too basic. We failed
to consider that a resource might be intentionally diskless, most
prominently when using DRBD's "quorum tiebreaker" feature. Up until now,
this would have been considered "bad", fix this to accept diskless
resources as "good".

This was exposed by the fact that LINSTOR now automatically deploys a
tiebreaker resource by default.