Added CRM template variable BracketedAllowedIPs
There are now two variants of the AllowedIPs variable:
  - AllowedIPs, which simply contains the IP/CIDR combination
  - BracketedAllowedIPs, which is identical to AllowedIPs for IPv4
    addresses, but contains a bracketed version, [IP/CIDR], of
    the same information for IPv6 addresses.

The reason for this change is that some resource agents, such as
the NFS export RA, do not work with the unbracketed format of IPv6
addresses, while the bracketed format may not work for other
purposes.
This change makes both versions available for CRM templates.