remove size argument from LUN
Turns out the size parameter isn't really needed in the iSCSI LUN. The
only place where it is actually used is when it is assigned to the
LINSTOR struct, and we can just assign that directly. The underlying
driver does not use any size argument.

Note that this changes the public REST API interface.