linstor: forbid existing resource in some situations
EnsureResource is built so that it does not error out when the resource
already exists. This is great for usability in general, but there is one
caveat:

We only use the very last part of the IQN/NQN for the LINSTOR resource
name, which means that one could create an iSCSI and an NVMe-oF target
with identical "unique" parts but different dates or domains. Gateway
would not detect this, since it is happy with the resource already
existing.

To fix this, require the resource to not exist when initially creating a
resource. This also has the bonus of making sure that the user does not
create a target with a "unique" name that corresponds to a LINSTOR
resource that already exists out of linstor-gateways control.