nfs: implement best practices
This implements the following recommendations that LINBIT usually
gives to customers implementing NFS HA clusters:

The first volume (#0) of any NFS resource is now reserved as a
"private cluster volume". This volume contains a filesystem with two
purposes. First, the NFS data directory (/var/lib/nfs) resides there,
making it a shared resource for the cluster. This ensures smooth
failovers, keeping client connections intact. The second purpose is
for the "tickle_dir". This serves the purpose of keeping TCP
connections alive during a failover.

Also, the nfs_server_scope parameter is now set for NFS resources. This,
again, is useful in a failover scenario.