reactor: switch to pelletier/go-toml
One problem with the generated drbd-reactor configurations is that the
toml marshaler we used puts all the "start" entries on one long line.
This makes the generated configuration almost impossible to read and
debug.

This new library supports putting every array element on a new line,
so use that instead.

NOTE: we cannot use v2 because it apparently does not support the
"UnmarshalTOML" method.