iscsi: set scsi_sn to a unique value
Turns out ESXi gets a bit confused when it sees multiple logical units
with the same SCSI serial number. This causes troubles on failover of
multiple targets at once.

The default is actually already supposed to be "reasonably unique", but
that uses a hash of the OCF resource name, which we always set to the
same value ("lu1").

Set the scsi_sn to an *actually* unique value to fix this -- use the IQN
for creating the hash.