add support for gross size
Most of the time, users will want to specify the net size for their
volumes. That is, the size that will actually be usable in the exported
volume later.

However, there are use cases where you would want to specify the gross
size; meaning the size that the volume including all metadata etc. will
occupy on the physical disk. The most common case is "use all available
space", where the user wants to utilize 100% of the backing disk,
regardless of how much net space this yields.

LINSTOR supports this via the GROSS_SIZE volume flag, so add a command
line flag to LINSTOR Gateway and pass it through.