update golinstor, fix volume number zero bug
Previously, when we specified volume number zero on a create request, it
would reach golinstor as a literal zero (which is the zero value for
int), so it would get omitted as empty when the JSON was serialized. To
avoid this problem, the volume number was converted to a pointer.