build: actually build program in lbbuild
For the longest time, we have done "fake" builds of linstor-gateway.
This means that the binaries are built on the packager's workstation and
included in the source tarball. lbbuild then just copies the binary over
to the packages.

This has several disadvantages; it makes cross-architecture builds
difficult and sets us up for CGO glibc incompatibilities, to name a few.

So, lbbuild was modified to actually build the go code. We can now
package the source files into the tarball and have lbbuild take care of
the building.