build: make Makefile usable from lbbuild
In the lbbuild environment, we do not have a git checkout available. But
when building for debian, we want to just call "make" and be done.
So encode everything in version.env when building debrelease (on the
development machine), then read the version and git hash from
version.env when building the binary.

Also include the Makefile in the tarball.