iscsi,nfs,nvmeof: fix nil deref on list
If the promoter config for a resource exists, but the LINSTOR resource
for some reason does not (for example, after a manual "linstor rd d"),
the "list" command would produce a nil dereference in the backend.

Fix this by just making the LINSTOR state "as bad as possible" if we
don't have a LINSTOR resource state.

Also give the user some guidance on what to do next (run "linstor v l"
and hopefully notice that the resource does not exist).