integration tests: rm iscsi-dirty-delete
That did not last long... Since we now check for IP collisions, this can
no longer be easily triggered. Since the collision check is run first,
Gateway will refuse to create the target the second time.

I think that this behavior is nicer anyway, so let's just remove this
test.