common: check that State is set before accessing
If one node where a resource is deployed is offline, its State will be
nil. Since we unconditionally access State, this will lead to a nil
pointer dereference.

Fix this by just ignoring the resource if its State is nil. This is fine
in AnyResourcesInUse because a resource on an offline node cannot
possibly be InUse.