linstor: roll back created resource definitions on failure
One of the most annoying things about linstor-gateway is this: say you
create a resource, but something goes wrong (for example, there is too
little space in your storage pools).
linstor-gateway will give an error, you fix it, and try again. And now
it fails because of a name collision because the resource definition
already exists from the previous run.

Mitigate some of that annoyance by trying to clean up after ourselves.
If the create operation fails, roll back by deleting the created
resource definition.