wait for resources to be running stably
Currently, we only check that the created resource becomes InUse on any
node. However, if the resource agent encounters an error on startup, it
will briefly become InUse on one node, but it will move away from that
node once startup fails.

This is especially confusing to users because linstor-gateway will show
the resource as "running" (which actually means something closer to
"deployed"), but in reality it is hopping between nodes, unable to
start.

So, when deploying a resource, wait a few seconds to see if the LINSTOR
resource stays on one node consistently. Also, if it does fail to
deploy, make sure to remove the reactor config file from LINSTOR's
database. That way, the creation process should be fully rolled back if
something goes wrong.

There are two scenarios that this does not cover:
* Deployment only fails on one node, then switches to another and
  succeeds. This would be seen as a failure to deploy.
* The resource agent takes longer than 5 seconds to fail.

Both are firmly in the "too bad" category, but we may still want to
extend the behavior in the future to cover these cases.