nfs: relax root_owner parsing
When we read a config back in, we want to parse the root_owner value
from the MkfsParams. Currently, this is done via Sscanf, which is
relatively strict and would break if we added more flags to MkfsParams.
Change the implementation to use a regex so that it matches even if we
add more flags.