linstor: add nodiscard flag to MkfsParams
Discard sizes in DRBD are relatively small, and discards always have an
IO depth of 1, so creating a file system on a large backing device might
take a long time. Disabling the use of discards during file system
creation fixes this.

Linstor already adds this flag by default, but since we overwrite the
MkfsParams value here we also need to add it manually.