linstor: use user and group name instead of id
Recent LINSTOR versions support passing the user id and group as names
instead of UID and GID. This means that we can pass "nobody:nobody"
directly instead of making a guess about what the UID and GID will be.

Try to be backward compatible by parsing back the UID and GID if the
MkfsParams property exists and passing the numbers back to LINSTOR as
user and group strings.