iscsi: make scsi_sn unique per volume
Some initiators do not like it if the scsi_sn of two LUs is the same
value. This makes sense, since the scsi_sn should uniquely identify a
single storage device.

So generate the scsi_sn in a way that makes it a unique value for
every volume of the created resource.

Also incorporate the scsi_sn into the product_id, since SCST passes
this value for the t10_dev_id parameter, which some initiators
(e.g. xenserver) factor into their own "unique ID" calculations.