iscsi: disallow uppercase characters in IQN
LIO converts all IQNs to lowercase when creating iSCSI resources. If
uppercase characters are present in the IQN, this will create a
mismatch (we expect the mixed-case version, LIO expects lowercase),
causing resource creation to fail.

To prevent this, tighten our rules for IQN names and disallow uppercase
characters entirely, and reflect this in the relevant error message.