iscsi: make first 8 bytes of product_id unique
VMware identifies iSCSI devices via their product_id, but for whatever
reason only looks at the first 8 bytes. We set the product_id to
"LINSTOR xxx", where xxx is a unique hash value. But the first 8
characters of that string are "LINSTOR ", which leads VMware to believe
that all LUNs are equivalent.

Remove the "LINSTOR " prefix to make the value unique.