server: change default port to 8337
In hindsight, it probably wasn't the best idea to choose this incredibly
common port for the REST API.
We have now encountered the first conflict in the real world, so change
the port.
LINSTOR uses port 3370, so choose 8337 to keep some association.

Luckily, the API probably does not have many consumers, so this should
go rather painlessly.